Move FlexDirection style tests to the flex style module
